requirements.txt
streamlit
requests
Pillow
langchain
langchain_community
